feat: add PhotoUploader component for enhanced photo management in ProcessOrder page, replace previous photo upload logic for improved user experience
