+++ src/routes/deliver.routes.tsx
@@ -4,7 +4,7 @@ import DeliverLayout from '../layouts/DeliverLayout';
 
 //外送員頁面組件
 import Task from '../pages/deliver/Task'; //任務清單
-import OrderDetailDeliver from '../pages/deliver/Task/OrderDetail'; //方案詳情
+import TaskDetail from '../pages/deliver/TaskDetail'; //方案詳情
 import ScanOrder from '../pages/deliver/ScanOrder'; //掃描訂單
 import ProcessOrder from '../pages/deliver/ProcessOrder'; //處理訂單
 import Calendar from '../pages/deliver/Calendar'; //行事曆
@@ -29,7 +29,7 @@ export const deliverRoutes: RouteConfig[] = [
     path: '/deliver/task/:taskId',
     element: (
       <ProtectedRoute role="deliver">
-        <OrderDetailDeliver />
+        <TaskDetail />
       </ProtectedRoute>
     ),
   },
