+++ src/styles/GlobalStyles.tsx
@@ -1,5 +1,7 @@
 import { createGlobalStyle } from 'styled-components';
 
+//推薦使用theme provider
+
 const GlobalStyles = createGlobalStyle`
 :root {
   /* Indigo */
