+++ src/types/deliver/TaskCard.ts
@@ -1,0 +1,13 @@
+import { TaskStatus } from './TaskStatus';
+
+export type TaskCardProps = {
+  taskId: string;
+  number: string;
+  status: TaskStatus;
+  time: string;
+  address: string;
+  notes: string;
+  onStatusChange?: (taskId: string, newStatus: TaskStatus) => void;
+  photos?: string[];
+  isDisabled?: boolean;
+};
