+++ src/types/deliver/TaskStatus.ts
@@ -1,10 +1,4 @@
 //後端回傳 為中文
-//異常
-//未排定
-//已排定
-//前往中
-//已抵達
-//已完成
 
 //前端轉換 為英文
 export type TaskStatus =
